finished

+++ mistakes_are_always_forgivable_if_one_has_the_courage_to_admit_them/2-calculator/get_op_func.c
@@ -1,6 +1,7 @@
 #include "calculator.h"
 
 int (*get_op_func(char c))(int, int) {
+  /*assigns different functions to different ops*/
   int (*func_ptr_array[256])(int, int);
   func_ptr_array[37] = op_mod;
   func_ptr_array[42] = op_mul;
@@ -8,8 +9,9 @@ int (*get_op_func(char c))(int, int) {
   func_ptr_array[45] = op_sub;
   func_ptr_array[47] = op_div;
 
+  /*error checking*/
   if (c != 37 && c != 42 && c != 43 && c != 45 && c != 47) {
     return 0;
   }
-  return func_ptr_array[c];
+  return (func_ptr_array[(int)c]);
 }
